Included a main module (SudokuPuzzleMain.py) and .gitignore.
Fleshed out the basic structure for the Sudoku solver. I opted for constraint propagation as the CSP search technique, and renamed some class methods in SudokuPuzzleSolver.py for better accuracy.

SudokuPuzzleSolver.py also detects the validity of a Sudoku board- that is, assuming a file gets corrupted, and/or I opt for a user inputting their own custom board.