Slightly fleshed out GUI. Added Button class.
Background is filled with gray, changed float calculations to be integers instead (due to deprecation warnings), and a pseudo-button is floating near the middle of the screen. Currently, only the text is seen, but the program is still able to detect the mouse hovering over it and subsequently register any clicks.