Working button hover. Both color and text update.
By clicking the button, you'll notice that the over-extended text bleeds behind the updated button. This is being temporarily kept for future optimization purposes.

This seems to be a common "glitch"- I've seen it occur in several other applications, but I still don't know the technical terminology to describe such a glitch.