Grid tiles are now detected and updatable.
Removed permanent storage of puzzle file and difficulty. They will be held in memory when they are needed.

Intentionally left in TODO item of "Better puzzle selection". For now, the very first puzzle in the puzzle package is selected for testing purposes.

Class Board's self.board is now a backed-list of button as opposed to a tuple. It would make more sense to have access to the button itself, as opposed to manually validating through a tuple of button positions. Currently, the program does "get_button_from_mouse_pos(self, mouse_pos)" (searches through each button manually) and determines if button.is_over(mouse_pos). Time cost is hopefully O(n) for 81 tiles. This is relatively small given this scope, but my biggest concern is potential sluggish speeds if not optimized properly.

The only bug found so far is that the "Play" button is drawn when clicking outside of a valid board tile.