Several GUI.py refactors. See notes below.
def initialize_board(self) now contains a volatile "offset". It's useful if you want something to be drawn not fully on screen, but still maintain a desired margin size. Directly adjusting self.margin to include the offset will absolutely alter the margin size in other functions that need it.

Furthermore, the method's set_submenu needs to have h redeclared, because it's y position depends on it. It just simply removes the offset.

Removed margin argument from def _set_tile_properties. The margin is already stored in self.margin.

Lastly, a submenu is drawn in the correct position, but currently does not contain any significant functionality.