Working implementation of side menu.
Side menu just has test text, but will be later implemented to include important information

Included a self.board_is_solved. This is to prevent the player from crashing the game if they solve a puzzle and then try clicking a tile- this is because that same board being access was just converted into a list of strings for validation purposes.

As a result of the above statement, a copy of the same was created to re-convert the board if the player fills up all of the tiles and it is not the correct solution.

Started puzzle index selector at -1. This is so the index automatically increments every time a new board is loaded.

Submenu text updating happens every frame as opposed to each individual mouse click- this means the original bug is fixed. Otherwise the elapsed text would only update every time the mouse moves or the player clicks.