Font accommodations based on user's Operating System.
Major refactoring (practically a rewrite) to configvideo.py. Due to having to process the operating system for valid fonts, it's "configuration" was transformed into a class. It's primary purpose is to search for the preferred font (the font that I've been testing with) in the user's operating system. If it does exist, it uses a "fallback font" that exists on Windows, Linux and MacOS. This is subject to change, but works nonetheless.

However, it currently only processes system fonts if the user's operating system is Windows 10. I have yet to explore other options on other Operating Systems.

In GUI.py, made some slight refactoring to accommodate the new changes in configvideo.py.

Also updated README.md.

I am leaving the below past commit message due it's extremely critical nature.
==============================================================================

Note: forgot to add pyinstaller compilation files to git. See previous commit notes below.

The most critical thing to keep in mind is that this project/program must use a valid system font (.ttf) in order to be successfully compiled into an .exe by pyinstaller https://pypi.org/project/PyInstaller/. This means you must use pygame's SysFont("name of font") and not Font("name of font.ttf"). Pygame's Font() function uses the package pkg_resource, and said package is not understood by pyinstaller.

Also, PyInstaller allegedly does not work with Python 3.7, so make sure your interpreter is using Python 3.6 when compiling.

The .exe itself must be in the same directory with the components/scripts that it was built with. This means the .exe needs to be manually moved out of the dist folder after it was compiled and moved into its parent directory.

As it currently stands, the .exe can be run for recreational purposes. Just make sure to download the full repository and move GUI.exe into the root / parent directory. This extra step exists for my own benefit- it's something that'll eventually require my attention and also indicates that this product is not yet ready for mass consumerism.

Lastly, all ' were changed to ". I initially assumed this was the cause of pyinstaller not successfully compiling GUI.py. Turns out there's no effect either way.