Restructuring in part due to major compatibility issues.
PyInstaller will package your projects if you are using PyGame's SysFont() function- with the caveat that your .exe will work half the time. An alternative solution was researched into. I discovered that I could simply use the Font() function and override it to use fonts that are downloaded somewhere in the project directory.

Such an override requires a string parameter of the root directory to the file name including extension. However, returning a string from another class will not work for reasons I currently don't understand. This means that fonts are essentially hard coded. For now, this is adequately sufficient but will still require research as to how I can pass in the directory in a format that Font() will understand.

configvideo.py now forces the user to return its font and size through methods, as opposed to accessing them directly. This seems to be the most intuitive way to override a font or font size on an individual basis. GUI.py should be more readability as a result of this change.

I have included the compilation files for future reference. Eventually, I'll have to remove the unused packages to cut down on file size. It's better if I remove them one-by-one and test it with pyinstaller's inherently finicky behavior.

README.md was also updated to reflect the superior nature of the development version of pyinstaller.