Full implementation of auto-solver and a bug fix.
Bug fix: less chance of crashes. It will still crash if you do something else while the program is running. Prevented inputs or any frame drawing while the computer is busy solving the board and updating the GUI.

===========================================================================
Older commit, still relevant:

New attribute for class Board called "self.board_is_dual_solving". Reasoning for naming convention: On the front-end of things, it looks like it's being automatically solved. Internally, however, is pseudo-dual solving. While the puzzle is being solved in the background, the GUI is being updating along side as a visual indicator of the back tracking algorithm. In other several areas, it helps to know when the GUI is "auto solving".

Edited method _process_solved_board(self). It no longer takes a parameter of a board. I had already stored the board inside "self".

Bug fix: IndexError when trying to access a puzzle higher than 46. The length of the puzzle is currently hard coded. I will later look into a more optimal solution for this bandaid fix.

Everything otherwise contains the same logistics as the last commit. The only differences are slight variable name changes and easier-to-read indentation.

Known issue: trying to close the window while the board is auto-solving will crash. Re-opening the window after the crash and trying to auto-solve will inevitably crash with no traceback.

When it comes to handling the solver class, things become very tricky. I want to maintain 100% source of the professor's original sudoku files- this means that there's back and forth data conversions and it's easy to get lost in def validate_solution(). It would absolutely be easier if the text files were formatted to my specific liking, but this is my decision after weighing the time-cost benefit. Current maintenance is also proving to be more challenging as the already-existing large code base becomes larger.